Rearrange codes
